src/conductor/requirements.txt
omxplayer-wrapper
RPI.GPIO
mutagen
pydub
numpy